pyproject.toml
[tool.poetry]
name = "tastytrade-sdk"
version = "0.0.31a0"
description = "A python wrapper around the tastytrade open API"
authors = ["Aaron Mamparo <aaronmamparo@gmail.com>"]
license = "MIT"
readme = "README.md"
packages = [
    { include = "tastytrade_sdk", from = "src" }
]

[project.urls]
documentation = "https://tastytrade-sdk.readthedocs.io"
repository = "https://github.com/amamparo/tastytrade-sdk-python"

[tool.poetry.dependencies]
python = "^3.8"
injector = "^0.20.1"
requests = "2.29.0"
websockets = "^11.0.3"
bidict = "^0.22.1"
strenum = "^0.4.15"
ujson = "^5.8.0"

[tool.poetry.group.dev.dependencies]
python-dotenv = "^1.0.0"
pylint = "^2.17.4"
pylint-quotes = "^0.2.3"
mkdocstrings = { extras = ["python"], version = "^0.21.2" }
mkdocs = "^1.4.3"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
